NYXIA CORREÇOES/requirements.txt
transformers
torch
langchain
python-dotenv
colorama
tqdm
textblob
rich
